LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/open-url-encoding-expected.txt


FAIL percent encode characters assert_equals: expected "%C3%9F" but got "%DF"
FAIL lone surrogate should return U+FFFD assert_equals: expected "%EF%BF%BD" but got "%26%2355357%3B"
